Reapply "🔥 댓글 추가 기능까지 완료"
This reverts commit bdb2db2bbf61a6a473e15c8ff6bf8e6db071d26f.
