+++ src/components/RestaurantMap.jsx
@@ -52,13 +52,14 @@ const RestaurantMap = () => {
 
       <MapContainer>
         <Map center={center} style={{ width: '100%', height: '100vh' }}>
-          {restaurants.map((restaurant) => (
-            <MapMarker
-              key={restaurant.id}
-              position={{ lat: restaurant.latitude, lng: restaurant.longitude }}
-              title={restaurant.name}
-            />
-          ))}
+          {restaurants &&
+            restaurants.map((restaurant) => (
+              <MapMarker
+                key={restaurant.id}
+                position={{ lat: restaurant.latitude, lng: restaurant.longitude }}
+                title={restaurant.name}
+              />
+            ))}
         </Map>
       </MapContainer>
     </Container>
